Added player movement controls!
I created a new file called Player.js where I created the player movement controls which are WASD for movement and Space for jumping.

I rotated the player in main.js to be facing the camera.

I made changes so the camera is focused on the player and will move with the player.